python-service/requirements.txt
pandas
sqlalchemy
openpyxl
psycopg2-binary
flask
python-dotenv

